requirements.txt
Flask==3.0.3
numpy==1.26.4
Pillow==10.3.0
Requests==2.32.2
tensorflow==2.16.1
tensorflow_intel==2.16.1
